feat(client): add shared insight id for panels in workspace
