+++ libs/renderer/src/Renderer.tsx
@@ -182,7 +182,8 @@ export const Renderer = observer((props: RendererProps) => {
                     background: "rgba(255, 255, 255, 0.5)",
                     zIndex: 1501,
                     position: "relative",
-                    width: "100%"
+                    width: "100%",
+                    height: "100%"
                 }}
             >
                 <Stack
@@ -191,6 +192,7 @@ export const Renderer = observer((props: RendererProps) => {
                     justifyContent={"center"}
                     spacing={1}
                     width={"100%"}
+                    height={"100%"}
                 >
                     <CircularProgress color={"info"} />
                 </Stack>
@@ -207,7 +209,8 @@ export const Renderer = observer((props: RendererProps) => {
                         background: "rgba(255, 255, 255, 0.5)",
                         zIndex: 1501,
                         position: "relative",
-                        width: "100%"
+                        width: "100%",
+                        height: "100%"
                     }}
                 >
                     <Stack
@@ -216,6 +219,7 @@ export const Renderer = observer((props: RendererProps) => {
                         justifyContent={"center"}
                         spacing={1}
                         width={"100%"}
+                        height={"100%"}
                     >
                         <CircularProgress color={"info"} />
                     </Stack>
