+++ packages/client/src/components/workspace/panels/TerminalPanel.tsx
@@ -19,6 +19,7 @@ import PythonLogo from '@/assets/img/Python-logo.svg';
 import RLogo from '@/assets/img/R-logo.svg';
 
 import { Panel } from './Panel';
+import { useWorkspace } from '@/hooks';
 
 const StyledImage = styled('img')(({ theme }) => ({
     height: '13px',
@@ -59,6 +60,7 @@ export const TerminalPanel: React.FC = observer(() => {
 
     const [history, setHistory] = useState<TerminalProps['history']>([]);
     const [isLoading, setIsLoading] = useState<boolean>(false);
+    const { workspace } = useWorkspace();
 
     const [command, setCommand] = useState<string>('');
     const [language, setLanguage] = useState('PIXEL');
@@ -108,8 +110,8 @@ export const TerminalPanel: React.FC = observer(() => {
             }
 
             // run the pixel
-            // TODO: Fix Insight ID
-            const response = await runPixel(pixel);
+            // TODO: We need to fix workspace.store so we just call runWorkspacePixel
+            const response = await workspace.runWorkspacePixel(pixel);
 
             const updatedHistory = [...history];
             for (const r of response.pixelReturn) {
